+++ lib/rspec/core/formatters/base_text_formatter.rb
@@ -31,16 +31,16 @@ module RSpec
         #
         # @param notification [NullNotification]
         def dump_failures(notification)
-          return if failed_example_notifications.empty?
+          return if notification.failure_notifications.empty?
           output.puts
           output.puts "Failures:"
-          failed_example_notifications.each_with_index do |failure, index|
+          notification.failure_notifications.each_with_index do |failure, index|
             output.puts
             output.puts "#{short_padding}#{index.next}) #{failure.description}"
-            failure.colorized_message_lines(ConsoleCodes).each do |line|
+            failure.colorized_message_lines.each do |line|
               output.puts "#{long_padding}#{line}"
             end
-            failure.colorized_formatted_backtrace(ConsoleCodes).each do |line|
+            failure.colorized_formatted_backtrace.each do |line|
               output.puts "#{long_padding}#{line}"
             end
           end
@@ -57,31 +57,18 @@ module RSpec
         def dump_summary(summary)
           output.puts "\nFinished in #{summary.formatted_duration}" +
                       " (files took #{summary.formatted_load_time} to load)\n"
-          output.puts summary.colorize_with ConsoleCodes
-          dump_commands_to_rerun_failed_examples
-        end
-
-        # @api public
-        #
-        # Outputs commands which can be used to re-run failed examples.
-        #
-        def dump_commands_to_rerun_failed_examples
-          return if failed_examples.empty?
-          output.puts
-          output.puts("Failed examples:")
-          output.puts
-
-          failed_examples.each do |example|
-            output.puts(failure_color("rspec #{RSpec::Core::Metadata::relative_path(example.location)}") + " " + detail_color("# #{example.full_description}"))
+          output.puts summary.colorized
+          unless summary.failed_examples.empty?
+            output.puts summary.colorized_rerun_commands
           end
         end
 
         # @private
         def dump_pending(notification)
-          unless pending_examples.empty?
+          unless notification.pending_examples.empty?
             output.puts
             output.puts "Pending:"
-            pending_examples.each do |pending_example|
+            notification.pending_examples.each do |pending_example|
               output.puts pending_color("  #{pending_example.full_description}")
               output.puts detail_color("    # #{pending_example.execution_result.pending_message}")
               output.puts detail_color("    # #{format_caller(pending_example.location)}")