+++ lib/rspec/core/formatters/documentation_formatter.rb
@@ -14,8 +14,6 @@ module RSpec
         end
 
         def example_group_started(notification)
-          super
-
           output.puts if @group_level == 0
           output.puts "#{current_indentation}#{notification.group.description.strip}"
 
@@ -31,12 +29,10 @@ module RSpec
         end
 
         def example_pending(pending)
-          super
           output.puts pending_output(pending.example, pending.example.execution_result.pending_message)
         end
 
         def example_failed(failure)
-          super
           output.puts failure_output(failure.example, failure.example.execution_result.exception)
         end